Added more verbose logging with performance profiling data
